Use CloudFront to serve images
Now all images are served from CloudFront rather than returning them
directly
in the lambda response body. Right now the downside to this is
all images are reachable at any point (previously the secretiveness
was kinda on purpose) but you'd have to store old URLs or guess new
ones right (unlikely).

This also sets the stage for a "weekly recap"