Change emoji options
Make it so there are two new emoji options and the previous
emoji options are marked as deprecated. This was done in such a way
that old counts could still be retrieved but new writes will fail for
"deprecated" reactions. Also changed the reaction images on the frontend
to be served via cloudfront instead of hardcoded base64 strings