Continue refactor of lambda-utils
Complete the Image DAOs. Weren't able to merge them into one
which is super frustrating but I couldn't find a good way to do it.
Left a todo about coming back to it later.
Next steps are adding to daily-setup-lambda a flag every five days
and if that flag is true to return the last five images to the frontend
on the get-image-lambda call.