Migrate to Single Table design
Start migrating to Single Table design. Move the image-info table first
by migrating the get image and set new image paths. Includes a glue script that
was set up manually in the Glue console. Since it was a one-off it wasn't worth writing
in CDK to deploy but want to track the code for posterity.

To get this script to work invovled setting up more AWS resources manually (glue database,
crawlers, etc.) but those won't be included.