Commiting for posteriy. Immediately going to unwind these chagnes
I don't think object lambda is the best way to do this. it is brutal to get the permissions
right and this is just way too complicated. Instead of going through all this trouble
I think I'm just going to do this with Lambda@Edge functions. It seems better supported
and on top of that it's easier to write CDK/actual lambda code for. Also easier to debug I think.
But I've never written a lambda@edge function before. Immediately I'm going to undo these commits