Move discord bot off of ec2 to fargate
Moved the discord bot from ec2 to fargate and in the process
started using sst. Wanted to mess around with sst v3 and I'd wanted
to move this to a container eventually anyways so that it was easier
to deploy (no more sshing yippee). I've deleted both the file and the stack
itself that previously containend the discord bot. Before doing so I made
an update to have the secret be retained on delete so that it stayed around.