feat: get keychain sharing working
Keychain sharing now works. Lots of cleanup to do before this is actually usuable but its
working which is cool! The key was the team-id is added to the prefix of whatever you put in
the xcode box. Since I was adding it as a prefix myself itw ould have needed to be at the front
twice. All the other react-native-keychain stuff was necessary though (I think?? never know for sure)