feat: Final share sheet commit
Now the share sheet uploads images to the initial image upload bucket.
There's very little verification of the images happening here so now I want to do that
with a background lambda function. That function will verify the contents,
make sure the user is authorized to upload for the provided group,
and then size the image accordingly.