feat: modify first rust lambda to work with sst
This is the first step to fully porting the app over to sst.
Moved the lambda for selecting images daily over first since
its the most self isolated. Also required setting up the ImageTable
that currently exists in CDK.

The end goal of this is to make it 1) super easy to have a dev stack
and a prod stack (in separate accounts) 2) have everything deploy
together using SST completely getting rid of CDK