refactor: move to Vite
WAs it worth it? I don't think so but I did it cause it seemed more ~modern~. oh well you live and you learn. Also changed production stacks to point
to prod.jtken.com to make DNS delegation easier. i think
there's still a route to handling dns in the prod account and
still hosting the webiste on jtken.com though. I'm still working through it atm and will hopefully be able to make it happen